Fixed C++ tests
Add OutputSchema::add_field method

Add set_stream_descriptor method and have it be responsible for clearing the column types map

Flip condition for checking if variant holds BitSetTag

Add tests for more complex ASTs

Homogenise error messages when invalid AST provided

Replace EXPECT_THROW with ASSERT_THROW

Test refactor
