+++ cpp/arcticdb/CMakeLists.txt
@@ -939,7 +939,9 @@ if(${TEST})
             processing/test/test_filter_and_project_sparse.cpp
             processing/test/test_type_promotion.cpp
             processing/test/test_operation_dispatch.cpp
-            processing/test/test_output_schema.cpp
+            processing/test/test_output_schema_aggregator_types.cpp
+            processing/test/test_output_schema_ast_validity.cpp
+            processing/test/test_output_schema_basic.cpp
             processing/test/test_parallel_processing.cpp
             processing/test/test_resample.cpp
             processing/test/test_set_membership.cpp
