+++ cpp/arcticdb/processing/clause.cpp
@@ -159,7 +159,7 @@ OutputSchema FilterClause::modify_schema(OutputSchema&& output_schema) const {
     check_column_presence(output_schema, *clause_info_.input_columns_, "Filter");
     auto root_expr = expression_context_->expression_nodes_.get_value(expression_context_->root_node_name_.value);
     std::variant<BitSetTag, DataType> return_type = root_expr->compute(*expression_context_, output_schema.column_types());
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(std::holds_alternative<BitSetTag>(return_type), "FilterClause AST produces a column, not a bitset");
+    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(std::holds_alternative<BitSetTag>(return_type), "FilterClause AST would produce a column, not a bitset");
     return output_schema;
 }
 
@@ -204,9 +204,8 @@ OutputSchema ProjectClause::modify_schema(OutputSchema&& output_schema) const {
     check_column_presence(output_schema, *clause_info_.input_columns_, "Project");
     auto root_expr = expression_context_->expression_nodes_.get_value(expression_context_->root_node_name_.value);
     std::variant<BitSetTag, DataType> return_type = root_expr->compute(*expression_context_, output_schema.column_types());
-    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(std::holds_alternative<DataType>(return_type), "ProjectClause AST produces a column, not a bitset");
-    output_schema.stream_descriptor_.add_scalar_field(std::get<DataType>(return_type), output_column_);
-    output_schema.column_types().emplace(output_column_, std::get<DataType>(return_type));
+    user_input::check<ErrorCode::E_INVALID_USER_ARGUMENT>(std::holds_alternative<DataType>(return_type), "ProjectClause AST would produce a bitset, not a column");
+    output_schema.add_field(output_column_, std::get<DataType>(return_type));
     return output_schema;
 }
 
@@ -474,22 +473,22 @@ std::vector<EntityId> AggregationClause::process(std::vector<EntityId>&& entity_
 
 OutputSchema AggregationClause::modify_schema(OutputSchema&& output_schema) const {
     check_column_presence(output_schema, *clause_info_.input_columns_, "Aggregation");
-    StreamDescriptor stream_desc(output_schema.stream_descriptor_.id());
-    stream_desc.add_field(output_schema.stream_descriptor_.field(*output_schema.stream_descriptor_.find_field(grouping_column_)));
+    const auto& input_stream_desc = output_schema.stream_descriptor();
+    StreamDescriptor stream_desc(input_stream_desc.id());
+    stream_desc.add_field(input_stream_desc.field(*input_stream_desc.find_field(grouping_column_)));
     stream_desc.set_index({0, IndexDescriptorImpl::Type::ROWCOUNT});
 
     for (const auto& agg: aggregators_){
         const auto& input_column_name = agg.get_input_column_name().value;
         const auto& output_column_name = agg.get_output_column_name().value;
-        const auto& input_column_type = output_schema.stream_descriptor_.field(*output_schema.stream_descriptor_.find_field(input_column_name)).type().data_type();
+        const auto& input_column_type = input_stream_desc.field(*input_stream_desc.find_field(input_column_name)).type().data_type();
         auto agg_data = agg.get_aggregator_data();
         agg_data.add_data_type(input_column_type);
         const auto& output_column_type = agg_data.get_output_data_type();
         stream_desc.add_scalar_field(output_column_type, output_column_name);
     }
 
-    output_schema.stream_descriptor_ = std::move(stream_desc);
-    output_schema.clear_column_types();
+    output_schema.set_stream_descriptor(std::move(stream_desc));
     auto mutable_index = output_schema.norm_metadata_.mutable_df()->mutable_common()->mutable_index();
     mutable_index->set_name(grouping_column_);
     mutable_index->clear_fake_name();
@@ -530,22 +529,22 @@ void ResampleClause<closed_boundary>::set_component_manager(std::shared_ptr<Comp
 
 template<ResampleBoundary closed_boundary>
 OutputSchema ResampleClause<closed_boundary>::modify_schema(OutputSchema&& output_schema) const {
-    check_is_timeseries(output_schema.stream_descriptor_, "Resample");
+    check_is_timeseries(output_schema.stream_descriptor(), "Resample");
     check_column_presence(output_schema, *clause_info_.input_columns_, "Resample");
-    StreamDescriptor stream_desc(output_schema.stream_descriptor_.id());
-    stream_desc.add_field(output_schema.stream_descriptor_.field(0));
+    const auto& input_stream_desc = output_schema.stream_descriptor();
+    StreamDescriptor stream_desc(input_stream_desc.id());
+    stream_desc.add_field(input_stream_desc.field(0));
     stream_desc.set_index(IndexDescriptorImpl(1, IndexDescriptor::Type::TIMESTAMP));
 
     for (const auto& agg: aggregators_){
         const auto& input_column_name = agg.get_input_column_name().value;
         const auto& output_column_name = agg.get_output_column_name().value;
-        const auto& input_column_type = output_schema.stream_descriptor_.field(*output_schema.stream_descriptor_.find_field(input_column_name)).type().data_type();
+        const auto& input_column_type = input_stream_desc.field(*input_stream_desc.find_field(input_column_name)).type().data_type();
         agg.check_aggregator_supported_with_data_type(input_column_type);
         auto output_column_type = agg.generate_output_data_type(input_column_type);
         stream_desc.add_scalar_field(output_column_type, output_column_name);
     }
-    output_schema.stream_descriptor_ = std::move(stream_desc);
-    output_schema.clear_column_types();
+    output_schema.set_stream_descriptor(std::move(stream_desc));
 
     if (output_schema.norm_metadata_.df().common().has_multi_index()) {
         const auto& multi_index = output_schema.norm_metadata_.mutable_df()->mutable_common()->multi_index();
@@ -1008,7 +1007,7 @@ const ClauseInfo& MergeClause::clause_info() const {
 }
 
 OutputSchema MergeClause::modify_schema(OutputSchema&& output_schema) const {
-    check_is_timeseries(output_schema.stream_descriptor_, "Merge");
+    check_is_timeseries(output_schema.stream_descriptor(), "Merge");
     return output_schema;
 }
 
@@ -1323,7 +1322,7 @@ std::vector<EntityId> DateRangeClause::process(std::vector<EntityId> &&entity_id
 }
 
 OutputSchema DateRangeClause::modify_schema(OutputSchema&& output_schema) const {
-    check_is_timeseries(output_schema.stream_descriptor_, "DateRange");
+    check_is_timeseries(output_schema.stream_descriptor(), "DateRange");
     return output_schema;
 }
